move cmd/util/ to internal/cmd/util/, since it is also imported by packages outside of cmd/
