+++ cmd/acmesolver/main.go
@@ -21,7 +21,7 @@ import (
 	"os"
 
 	"github.com/cert-manager/cert-manager/cmd/acmesolver/app"
-	"github.com/cert-manager/cert-manager/cmd/util"
+	"github.com/cert-manager/cert-manager/internal/cmd/util"
 )
 
 // acmesolver solves ACME http-01 challenges. This is intended to run as a pod
