+++ cmd/controller/main.go
@@ -20,7 +20,7 @@ import (
 	"flag"
 
 	"github.com/cert-manager/cert-manager/cmd/controller/app"
-	"github.com/cert-manager/cert-manager/cmd/util"
+	"github.com/cert-manager/cert-manager/internal/cmd/util"
 	logf "github.com/cert-manager/cert-manager/pkg/logs"
 )
