+++ cmd/ctl/pkg/check/api/api.go
@@ -32,7 +32,7 @@ import (
 	"k8s.io/kubectl/pkg/util/templates"
 
 	"github.com/cert-manager/cert-manager/cmd/ctl/pkg/factory"
-	cmcmdutil "github.com/cert-manager/cert-manager/cmd/util"
+	cmcmdutil "github.com/cert-manager/cert-manager/internal/cmd/util"
 	"github.com/cert-manager/cert-manager/pkg/util/cmapichecker"
 )
