+++ cmd/webhook/main.go
@@ -19,8 +19,8 @@ package main
 import (
 	"flag"
 
-	"github.com/cert-manager/cert-manager/cmd/util"
 	"github.com/cert-manager/cert-manager/cmd/webhook/app"
+	"github.com/cert-manager/cert-manager/internal/cmd/util"
 	logf "github.com/cert-manager/cert-manager/pkg/logs"
 )
