+++ pkg/acme/webhook/cmd/cmd.go
@@ -23,7 +23,7 @@ import (
 
 	"k8s.io/component-base/logs"
 
-	"github.com/cert-manager/cert-manager/cmd/util"
+	"github.com/cert-manager/cert-manager/internal/cmd/util"
 	"github.com/cert-manager/cert-manager/pkg/acme/webhook"
 	"github.com/cert-manager/cert-manager/pkg/acme/webhook/cmd/server"
 	logf "github.com/cert-manager/cert-manager/pkg/logs"
